requirements/base.txt
Django==3.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
dj_database_url==0.5.0
psycopg2-binary==2.9.3
django-cors-headers==3.2.0
